requirements.txt
flask
flask-cors
gunicorn
gspread
google-auth
pandas
openpyxl
requests
urllib3